+++ graphql/schema.graphql
@@ -26,3 +26,4 @@ scalar Null  @scalar(class: "MLL\\GraphQLScalars\\NullScalar")
 #import schemas/Social/*.graphql
 #import schemas/ContentEngine/*.graphql
 #import schemas/Souk/*.graphql
+#import schemas/Ecosystem/Admin/*.graphql
