+++ src/Kanvas/AccessControlList/Enums/RolesEnums.php
@@ -10,6 +10,7 @@ use Kanvas\Enums\AppEnums;
 
 enum RolesEnums: string
 {
+    case OWNER = 'Owner';
     case ADMIN = 'Admins';
     case USER = 'Users';
     case AGENT = 'Agents';
