+++ src/Kanvas/Auth/Actions/RegisterUsersAppAction.php
@@ -26,6 +26,8 @@ class RegisterUsersAppAction
 
     /**
      * Register an user into a new app with a password for the login.
+     * 
+     * @psalm-suppress MixedReturnStatement
      */
     public function execute(string $password): UsersAssociatedApps
     {
