+++ tests/GraphQL/Ecosystem/Apps/UserManagementTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\GraphQL\Ecosystem\Apps;
 
+use Kanvas\AccessControlList\Enums\RolesEnums;
 use Kanvas\Apps\Models\Apps;
 use Kanvas\Auth\Actions\RegisterUsersAppAction;
 use Kanvas\Enums\AppEnums;
@@ -150,6 +151,43 @@ class UserManagementTest extends TestCase
         $response->assertJson([
             'data' => [
                 'appUserUpdateEmail' => true,
+            ],
+        ]);
+    }
+
+    public function testCreateUser()
+    {
+        $app = app(Apps::class);
+
+        $user = auth()->user();
+        $user->assign(RolesEnums::ADMIN->value);
+        
+        $email = fake()->email();
+        $response = $this->graphQL(/** @lang GraphQL */ '
+            mutation appCreateUser($data: CreateUserInput!) {
+                appCreateUser(data: $data) {
+                    id
+                    email
+                }
+              }',
+            [
+                'data' => [
+                    'firstname' => fake()->firstName(),
+                    'lastname' => fake()->lastName(),
+                    'email' => $email,
+                ],
+            ],
+            [],
+            [
+                AppEnums::KANVAS_APP_KEY_HEADER->getValue() => $app->keys()->first()->client_secret_id,
+            ]
+        );
+
+        $response->assertJson([
+            'data' => [
+                'appCreateUser' => [
+                    'email' => $email,
+                ],
             ],
         ]);
     }
